+++ MelonLoader/Core.cs
@@ -10,24 +10,24 @@ namespace MelonLoader
 {
     internal static class Core
     {
+        internal static string GameDir = null;
         internal static string UserDataPath = null;
 
         static Core()
         {
-            string gamedir = MelonUtils.GetGameDirectory();
-            ((AppDomainSetup)typeof(AppDomain).GetProperty("SetupInformationNoCopy", BindingFlags.NonPublic | BindingFlags.Instance).GetValue(AppDomain.CurrentDomain, new object[0])).ApplicationBase = gamedir;
-            Directory.SetCurrentDirectory(gamedir);
+            GameDir = MelonUtils.GetGameDirectory();
+            ((AppDomainSetup)typeof(AppDomain).GetProperty("SetupInformationNoCopy", BindingFlags.NonPublic | BindingFlags.Instance).GetValue(AppDomain.CurrentDomain, new object[0])).ApplicationBase = GameDir;
+            Directory.SetCurrentDirectory(GameDir);
             AppDomain.CurrentDomain.UnhandledException += UnhandledExceptionHandler;
             AppDomain.CurrentDomain.AssemblyResolve += AssemblyResolveHandler;
             AppDomain.CurrentDomain.ReflectionOnlyAssemblyResolve += AssemblyResolveHandler;
             CurrentCultureFix();
-            UserDataPath = Path.Combine(gamedir, "UserData");
+            UserDataPath = Path.Combine(GameDir, "UserData");
             if (!Directory.Exists(UserDataPath))
                 Directory.CreateDirectory(UserDataPath);
-            MelonPreferences.LegacyCheck();
-            MelonPreferences.Load();
-            if (MelonPreferences.WasLegacyLoaded)
-                MelonPreferences.Save();
+            try { MelonPreferences.LegacyCheck(); } catch (Exception ex) { MelonLogger.Error("MelonPreferences.LegacyCheck Exception: " + ex.ToString()); throw ex; }
+            try { MelonPreferences.Load(); } catch (Exception ex) { MelonLogger.Error("MelonPreferences.Load Exception: " + ex.ToString()); throw ex; }
+            if (MelonPreferences.WasLegacyLoaded) try { MelonPreferences.Save(); } catch (Exception ex) { MelonLogger.Error("MelonPreferences.Save Exception: " + ex.ToString()); throw ex; } 
         }
 
         private static void Initialize()
@@ -85,7 +85,6 @@ namespace MelonLoader
         private static void CurrentCultureFix()
         {
             Harmony.HarmonyInstance harmonyInstance = Harmony.HarmonyInstance.Create("CurrentCultureFix");
-            harmonyInstance.Patch(typeof(Thread).GetProperty("CurrentCulture", BindingFlags.Public | BindingFlags.Instance).GetGetMethod(), new Harmony.HarmonyMethod(typeof(Core).GetMethod("GetCurrentCulturePrefix", BindingFlags.NonPublic | BindingFlags.Static)));
             try { harmonyInstance.Patch(typeof(Thread).GetProperty("CurrentCulture", BindingFlags.Public | BindingFlags.Instance).GetGetMethod(), new Harmony.HarmonyMethod(typeof(Core).GetMethod("GetCurrentCulturePrefix", BindingFlags.NonPublic | BindingFlags.Static))); } catch (Exception ex) {}
             try { harmonyInstance.Patch(typeof(Thread).GetProperty("CurrentUICulture", BindingFlags.Public | BindingFlags.Instance).GetGetMethod(), new Harmony.HarmonyMethod(typeof(Core).GetMethod("GetCurrentCulturePrefix", BindingFlags.NonPublic | BindingFlags.Static))); } catch (Exception ex) {}
         }